added kappa4Kaufmann
Added a new function for calculating kappa4 from the relation in Kaufmann and Dreybrodt (2007). I could now add a pwpRateKaufmann, but I need to figure out how this kappa4 was determined, and how it should really be used to calculate the total rate (i.e. do I use equilibrium concentrations as in pwpTheory?). I also fixed some errors in units of kappa in the documentation.